factorial tests
