+++ Solvers.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+# Solver class provides methods for solving mathematical problems
+class Solver
+
+end
